FAM/requirements.txt
fastapi
uvicorn[standard]
pymongo
pymongo[srv]
motor
pydantic[email]
passlib[bcrypt]
fastapi-mail
python-multipart
python-jose[cryptography]